refactor(chatd): extract relay/pubsub merge logic behind SubscribeFn
Move all multi-replica coordination code from the OSS chatd package
to the enterprise package behind a single SubscribeFn interface.

The OSS Subscribe method is now ~40 lines of local-only forwarding.
When SubscribeFn is nil (AGPL single-instance), Subscribe uses a
simple goroutine that pipes localParts to the output channel.

Enterprise sets SubscribeFn via newMultiReplicaSubscribeFn() which
contains the full relay lifecycle (openRelayAsync, closeRelay,
scheduleRelayReconnect), pubsub notification handling, reconnect
timers, and stale-dial detection.

Changes:
- chatd: Replace RemotePartsProvider + inline relay code with
  SubscribeFn type and SubscribeMultiReplicaParams struct
- chatd: Subscribe delegates to subscribeFn when non-nil
- enterprise/coderd: New chatsubscribe.go with all relay logic
- enterprise/coderd: remotePartsProvider type moved from chatd
  (now unexported, enterprise-internal)
- enterprise/coderd: coderd.go wires newMultiReplicaSubscribeFn
- coderd: Options.ChatRemotePartsProvider -> ChatSubscribeFn
- Tests: relay tests moved to enterprise/coderd/
- Fix goroutine leak in wrappedParts range loop (pre-existing)
