+++ coderd/chatd/chatd_test.go
@@ -6,7 +6,6 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"net/http"
 	"strings"
 	"sync"
 	"sync/atomic"
@@ -28,7 +27,6 @@ import (
 	"github.com/coder/coder/v2/coderd/database/dbgen"
 	"github.com/coder/coder/v2/coderd/database/dbtestutil"
 	dbpubsub "github.com/coder/coder/v2/coderd/database/pubsub"
-	coderdpubsub "github.com/coder/coder/v2/coderd/pubsub"
 	"github.com/coder/coder/v2/coderd/util/slice"
 	"github.com/coder/coder/v2/codersdk"
 	"github.com/coder/coder/v2/provisioner/echo"
@@ -1133,29 +1131,6 @@ func newTestServer(
 	return server
 }
 
-func newTestServerWithRelay(
-	t *testing.T,
-	db database.Store,
-	ps dbpubsub.Pubsub,
-	replicaID uuid.UUID,
-	provider chatd.RemotePartsProvider,
-) *chatd.Server {
-	t.Helper()
-
-	logger := slogtest.Make(t, &slogtest.Options{IgnoreErrors: true})
-	server := chatd.New(chatd.Config{
-		Logger:                     logger,
-		Database:                   db,
-		ReplicaID:                  replicaID,
-		Pubsub:                     ps,
-		RemotePartsProvider:        provider,
-		PendingChatAcquireInterval: testutil.WaitSuperLong,
-	})
-	t.Cleanup(func() {
-		require.NoError(t, server.Close())
-	})
-	return server
-}
 
 func seedChatDependencies(
 	ctx context.Context,
@@ -1213,292 +1188,7 @@ func setOpenAIProviderBaseURL(
 	require.NoError(t, err)
 }
 
-func TestSubscribeRelayReconnectsOnDrop(t *testing.T) {
-	t.Parallel()
 
-	db, ps := dbtestutil.NewDB(t)
-	workerID := uuid.New()
-	subscriberID := uuid.New()
-
-	var callCount atomic.Int32
-
-	provider := func(ctx context.Context, _ uuid.UUID, _ uuid.UUID, _ http.Header) (
-		[]codersdk.ChatStreamEvent, <-chan codersdk.ChatStreamEvent, func(), error,
-	) {
-		call := callCount.Add(1)
-		ch := make(chan codersdk.ChatStreamEvent, 10)
-		if call == 1 {
-			// First relay: send a part then close to simulate a drop.
-			ch <- codersdk.ChatStreamEvent{
-				Type: codersdk.ChatStreamEventTypeMessagePart,
-				MessagePart: &codersdk.ChatStreamMessagePart{
-					Role: "assistant",
-					Part: codersdk.ChatMessagePart{Type: codersdk.ChatMessagePartTypeText, Text: "first-relay"},
-				},
-			}
-			close(ch)
-		} else {
-			// Second relay: send a different part, keep open.
-			ch <- codersdk.ChatStreamEvent{
-				Type: codersdk.ChatStreamEventTypeMessagePart,
-				MessagePart: &codersdk.ChatStreamMessagePart{
-					Role: "assistant",
-					Part: codersdk.ChatMessagePart{Type: codersdk.ChatMessagePartTypeText, Text: "second-relay"},
-				},
-			}
-			// Don't close — keep alive so the subscriber stays connected.
-		}
-		return nil, ch, func() {}, nil
-	}
-
-	subscriber := newTestServerWithRelay(t, db, ps, subscriberID, provider)
-
-	ctx := testutil.Context(t, testutil.WaitLong)
-	user, model := seedChatDependencies(ctx, t, db)
-
-	// Create a chat and mark it as running on a remote worker.
-	chat, err := subscriber.CreateChat(ctx, chatd.CreateOptions{
-		OwnerID:            user.ID,
-		Title:              "relay-reconnect",
-		ModelConfigID:      model.ID,
-		InitialUserContent: []fantasy.Content{fantasy.TextContent{Text: "hello"}},
-	})
-	require.NoError(t, err)
-
-	chat, err = db.UpdateChatStatus(ctx, database.UpdateChatStatusParams{
-		ID:          chat.ID,
-		Status:      database.ChatStatusRunning,
-		WorkerID:    uuid.NullUUID{UUID: workerID, Valid: true},
-		StartedAt:   sql.NullTime{Time: time.Now(), Valid: true},
-		HeartbeatAt: sql.NullTime{Time: time.Now(), Valid: true},
-	})
-	require.NoError(t, err)
-
-	_, events, cancel, ok := subscriber.Subscribe(ctx, chat.ID, nil, 0)
-	require.True(t, ok)
-	t.Cleanup(cancel)
-
-	// Should get the first relay part.
-	require.Eventually(t, func() bool {
-		select {
-		case event := <-events:
-			if event.Type == codersdk.ChatStreamEventTypeMessagePart &&
-				event.MessagePart != nil &&
-				event.MessagePart.Part.Text == "first-relay" {
-				return true
-			}
-			return false
-		default:
-			return false
-		}
-	}, testutil.WaitMedium, testutil.IntervalFast)
-
-	// After the first relay closes, a reconnection should happen and
-	// deliver the second relay part.
-	require.Eventually(t, func() bool {
-		select {
-		case event := <-events:
-			if event.Type == codersdk.ChatStreamEventTypeMessagePart &&
-				event.MessagePart != nil &&
-				event.MessagePart.Part.Text == "second-relay" {
-				return true
-			}
-			return false
-		default:
-			return false
-		}
-	}, testutil.WaitMedium, testutil.IntervalFast)
-
-	require.GreaterOrEqual(t, int(callCount.Load()), 2)
-}
-
-func TestSubscribeRelayAsyncDoesNotBlock(t *testing.T) {
-	t.Parallel()
-
-	db, ps := dbtestutil.NewDB(t)
-	workerID := uuid.New()
-	subscriberID := uuid.New()
-
-	dialStarted := make(chan struct{})
-	dialContinue := make(chan struct{})
-
-	provider := func(ctx context.Context, _ uuid.UUID, _ uuid.UUID, _ http.Header) (
-		[]codersdk.ChatStreamEvent, <-chan codersdk.ChatStreamEvent, func(), error,
-	) {
-		// Signal that the dial has started, then block until released.
-		select {
-		case <-dialStarted:
-		default:
-			close(dialStarted)
-		}
-		select {
-		case <-dialContinue:
-		case <-ctx.Done():
-			return nil, nil, nil, ctx.Err()
-		}
-		ch := make(chan codersdk.ChatStreamEvent, 10)
-		return nil, ch, func() {}, nil
-	}
-
-	subscriber := newTestServerWithRelay(t, db, ps, subscriberID, provider)
-
-	ctx := testutil.Context(t, testutil.WaitLong)
-	user, model := seedChatDependencies(ctx, t, db)
-
-	// Create a chat in pending status.
-	chat, err := subscriber.CreateChat(ctx, chatd.CreateOptions{
-		OwnerID:            user.ID,
-		Title:              "relay-async-nonblock",
-		ModelConfigID:      model.ID,
-		InitialUserContent: []fantasy.Content{fantasy.TextContent{Text: "hello"}},
-	})
-	require.NoError(t, err)
-
-	// Subscribe before the chat is marked running so the relay opens
-	// via pubsub notification (openRelayAsync path).
-	_, events, cancel, ok := subscriber.Subscribe(ctx, chat.ID, nil, 0)
-	require.True(t, ok)
-	t.Cleanup(cancel)
-
-	// Now mark the chat as running on a remote worker. This publishes
-	// a status notification which triggers openRelayAsync on the
-	// subscriber.
-	notify := coderdpubsub.ChatStreamNotifyMessage{
-		Status:   string(database.ChatStatusRunning),
-		WorkerID: workerID.String(),
-	}
-	payload, err := json.Marshal(notify)
-	require.NoError(t, err)
-	err = ps.Publish(coderdpubsub.ChatStreamNotifyChannel(chat.ID), payload)
-	require.NoError(t, err)
-
-	// Wait for the relay dial to actually start (blocking in the
-	// provider).
-	select {
-	case <-dialStarted:
-	case <-ctx.Done():
-		t.Fatal("timed out waiting for relay dial to start")
-	}
-
-	// While the relay is still dialing (provider is blocked), publish
-	// another status change. If openRelayAsync blocked the select loop
-	// this event would never arrive.
-	statusNotify := coderdpubsub.ChatStreamNotifyMessage{
-		Status: string(database.ChatStatusWaiting),
-	}
-	statusPayload, err := json.Marshal(statusNotify)
-	require.NoError(t, err)
-	err = ps.Publish(coderdpubsub.ChatStreamNotifyChannel(chat.ID), statusPayload)
-	require.NoError(t, err)
-
-	// The waiting status event should arrive promptly despite the
-	// relay still dialing.
-	require.Eventually(t, func() bool {
-		select {
-		case event := <-events:
-			return event.Type == codersdk.ChatStreamEventTypeStatus &&
-				event.Status != nil &&
-				event.Status.Status == codersdk.ChatStatusWaiting
-		default:
-			return false
-		}
-	}, testutil.WaitShort, testutil.IntervalFast)
-
-	// Unblock the relay dial so the test can clean up.
-	close(dialContinue)
-}
-
-func TestSubscribeRelaySnapshotDelivered(t *testing.T) {
-	t.Parallel()
-
-	db, ps := dbtestutil.NewDB(t)
-	workerID := uuid.New()
-	subscriberID := uuid.New()
-
-	provider := func(_ context.Context, _ uuid.UUID, _ uuid.UUID, _ http.Header) (
-		[]codersdk.ChatStreamEvent, <-chan codersdk.ChatStreamEvent, func(), error,
-	) {
-		// Return a non-empty snapshot with two parts.
-		snapshot := []codersdk.ChatStreamEvent{
-			{
-				Type: codersdk.ChatStreamEventTypeMessagePart,
-				MessagePart: &codersdk.ChatStreamMessagePart{
-					Role: "assistant",
-					Part: codersdk.ChatMessagePart{Type: codersdk.ChatMessagePartTypeText, Text: "snap-one"},
-				},
-			},
-			{
-				Type: codersdk.ChatStreamEventTypeMessagePart,
-				MessagePart: &codersdk.ChatStreamMessagePart{
-					Role: "assistant",
-					Part: codersdk.ChatMessagePart{Type: codersdk.ChatMessagePartTypeText, Text: "snap-two"},
-				},
-			},
-		}
-		ch := make(chan codersdk.ChatStreamEvent, 10)
-		// Also send a live part after the snapshot.
-		ch <- codersdk.ChatStreamEvent{
-			Type: codersdk.ChatStreamEventTypeMessagePart,
-			MessagePart: &codersdk.ChatStreamMessagePart{
-				Role: "assistant",
-				Part: codersdk.ChatMessagePart{Type: codersdk.ChatMessagePartTypeText, Text: "live-part"},
-			},
-		}
-		return snapshot, ch, func() {}, nil
-	}
-
-	subscriber := newTestServerWithRelay(t, db, ps, subscriberID, provider)
-
-	ctx := testutil.Context(t, testutil.WaitLong)
-	user, model := seedChatDependencies(ctx, t, db)
-
-	// Create a chat already running on a remote worker.
-	chat, err := subscriber.CreateChat(ctx, chatd.CreateOptions{
-		OwnerID:            user.ID,
-		Title:              "relay-snapshot",
-		ModelConfigID:      model.ID,
-		InitialUserContent: []fantasy.Content{fantasy.TextContent{Text: "hello"}},
-	})
-	require.NoError(t, err)
-
-	_, err = db.UpdateChatStatus(ctx, database.UpdateChatStatusParams{
-		ID:          chat.ID,
-		Status:      database.ChatStatusRunning,
-		WorkerID:    uuid.NullUUID{UUID: workerID, Valid: true},
-		StartedAt:   sql.NullTime{Time: time.Now(), Valid: true},
-		HeartbeatAt: sql.NullTime{Time: time.Now(), Valid: true},
-	})
-	require.NoError(t, err)
-
-	initialSnapshot, events, cancel, ok := subscriber.Subscribe(ctx, chat.ID, nil, 0)
-	require.True(t, ok)
-	t.Cleanup(cancel)
-
-	// The initial snapshot should contain the two relay snapshot parts.
-	var snapshotTexts []string
-	for _, event := range initialSnapshot {
-		if event.Type == codersdk.ChatStreamEventTypeMessagePart && event.MessagePart != nil {
-			snapshotTexts = append(snapshotTexts, event.MessagePart.Part.Text)
-		}
-	}
-	require.Contains(t, snapshotTexts, "snap-one")
-	require.Contains(t, snapshotTexts, "snap-two")
-
-	// The live part should arrive on the events channel.
-	require.Eventually(t, func() bool {
-		select {
-		case event := <-events:
-			if event.Type == codersdk.ChatStreamEventTypeMessagePart &&
-				event.MessagePart != nil &&
-				event.MessagePart.Part.Text == "live-part" {
-				return true
-			}
-			return false
-		default:
-			return false
-		}
-	}, testutil.WaitMedium, testutil.IntervalFast)
-}
 
 func TestCloseDuringShutdownContextCanceledShouldRetryOnNewReplica(t *testing.T) {
 	t.Parallel()
@@ -1604,184 +1294,5 @@ func TestCloseDuringShutdownContextCanceledShouldRetryOnNewReplica(t *testing.T)
 		return fromDB.Status == database.ChatStatusWaiting &&
 			!fromDB.WorkerID.Valid &&
 			!fromDB.LastError.Valid
-	}, testutil.WaitMedium, testutil.IntervalFast)
-}
-
-// TestSubscribeRelayStaleDialDiscardedAfterInterrupt verifies that when a
-// user interrupts a streaming chat and sends a new message (which gets
-// picked up by a different replica), an in-flight relay dial to the
-// OLD replica is cancelled/discarded and the relay connects to the
-// NEW replica correctly.
-func TestSubscribeRelayStaleDialDiscardedAfterInterrupt(t *testing.T) {
-	t.Parallel()
-
-	db, ps := dbtestutil.NewDB(t)
-	oldWorkerID := uuid.New()
-	newWorkerID := uuid.New()
-	subscriberID := uuid.New()
-
-	// Gate to hold the first dial until we're ready.
-	firstDialStarted := make(chan struct{})
-	releaseFirstDial := make(chan struct{})
-
-	var callCount atomic.Int32
-
-	provider := func(ctx context.Context, _ uuid.UUID, workerID uuid.UUID, _ http.Header) (
-		[]codersdk.ChatStreamEvent, <-chan codersdk.ChatStreamEvent, func(), error,
-	) {
-		call := callCount.Add(1)
-		ch := make(chan codersdk.ChatStreamEvent, 10)
-		if call == 1 {
-			// First dial (to old worker): signal that we started,
-			// then block until released or context cancelled.
-			close(firstDialStarted)
-			select {
-			case <-releaseFirstDial:
-			case <-ctx.Done():
-				return nil, nil, nil, ctx.Err()
-			}
-			// If we get here after being released (not cancelled),
-			// return a stale part — this should be discarded.
-			ch <- codersdk.ChatStreamEvent{
-				Type: codersdk.ChatStreamEventTypeMessagePart,
-				MessagePart: &codersdk.ChatStreamMessagePart{
-					Role: "assistant",
-					Part: codersdk.ChatMessagePart{Type: codersdk.ChatMessagePartTypeText, Text: "stale-part"},
-				},
-			}
-			close(ch)
-			return nil, ch, func() {}, nil
-		}
-		// Second dial (to new worker): return a valid part.
-		ch <- codersdk.ChatStreamEvent{
-			Type: codersdk.ChatStreamEventTypeMessagePart,
-			MessagePart: &codersdk.ChatStreamMessagePart{
-				Role: "assistant",
-				Part: codersdk.ChatMessagePart{Type: codersdk.ChatMessagePartTypeText, Text: "new-worker-part"},
-			},
-		}
-		return nil, ch, func() {}, nil
-	}
-
-	subscriber := newTestServerWithRelay(t, db, ps, subscriberID, provider)
-
-	ctx := testutil.Context(t, testutil.WaitLong)
-	user, model := seedChatDependencies(ctx, t, db)
-
-	chat, err := subscriber.CreateChat(ctx, chatd.CreateOptions{
-		OwnerID:            user.ID,
-		Title:              "stale-dial-test",
-		ModelConfigID:      model.ID,
-		InitialUserContent: []fantasy.Content{fantasy.TextContent{Text: "hello"}},
-	})
-	require.NoError(t, err)
-
-	// Start chat in waiting state so Subscribe does NOT try an initial relay.
-	_, err = db.UpdateChatStatus(ctx, database.UpdateChatStatusParams{
-		ID:     chat.ID,
-		Status: database.ChatStatusWaiting,
-	})
-	require.NoError(t, err)
-
-	// Subscribe while chat is in "waiting" state — no relay opened.
-	_, events, cancel, ok := subscriber.Subscribe(ctx, chat.ID, nil, 0)
-	require.True(t, ok)
-	t.Cleanup(cancel)
-
-	// Now simulate the chat being picked up by the OLD worker via pubsub.
-	// This triggers openRelayAsync in the merge loop.
-	_, err = db.UpdateChatStatus(ctx, database.UpdateChatStatusParams{
-		ID:          chat.ID,
-		Status:      database.ChatStatusRunning,
-		WorkerID:    uuid.NullUUID{UUID: oldWorkerID, Valid: true},
-		StartedAt:   sql.NullTime{Time: time.Now(), Valid: true},
-		HeartbeatAt: sql.NullTime{Time: time.Now(), Valid: true},
-	})
-	require.NoError(t, err)
-	oldRunningNotify := coderdpubsub.ChatStreamNotifyMessage{
-		Status:   string(database.ChatStatusRunning),
-		WorkerID: oldWorkerID.String(),
-	}
-	oldRunningPayload, err := json.Marshal(oldRunningNotify)
-	require.NoError(t, err)
-	err = ps.Publish(coderdpubsub.ChatStreamNotifyChannel(chat.ID), oldRunningPayload)
-	require.NoError(t, err)
-
-	// Wait for the first dial goroutine to start (it's blocked in the provider).
-	select {
-	case <-firstDialStarted:
-	case <-ctx.Done():
-		t.Fatal("timed out waiting for first dial to start")
-	}
-
-	// Simulate interrupt: chat goes to "waiting".
-	_, err = db.UpdateChatStatus(ctx, database.UpdateChatStatusParams{
-		ID:     chat.ID,
-		Status: database.ChatStatusWaiting,
-	})
-	require.NoError(t, err)
-	waitingNotify := coderdpubsub.ChatStreamNotifyMessage{
-		Status: string(database.ChatStatusWaiting),
-	}
-	waitingPayload, err := json.Marshal(waitingNotify)
-	require.NoError(t, err)
-	err = ps.Publish(coderdpubsub.ChatStreamNotifyChannel(chat.ID), waitingPayload)
-	require.NoError(t, err)
-
-	// Wait for the merge loop to process the waiting notification
-	// and emit the status event before publishing the new running
-	// notification. This avoids time.Sleep (banned by project
-	// policy) and provides a deterministic sync point.
-	require.Eventually(t, func() bool {
-		select {
-		case event := <-events:
-			return event.Type == codersdk.ChatStreamEventTypeStatus &&
-				event.Status != nil &&
-				event.Status.Status == codersdk.ChatStatusWaiting
-		default:
-			return false
-		}
-	}, testutil.WaitMedium, testutil.IntervalFast)
-
-	// Now the chat transitions to running on the NEW worker.
-	_, err = db.UpdateChatStatus(ctx, database.UpdateChatStatusParams{
-		ID:          chat.ID,
-		Status:      database.ChatStatusRunning,
-		WorkerID:    uuid.NullUUID{UUID: newWorkerID, Valid: true},
-		StartedAt:   sql.NullTime{Time: time.Now(), Valid: true},
-		HeartbeatAt: sql.NullTime{Time: time.Now(), Valid: true},
-	})
-	require.NoError(t, err)
-	runningNotify := coderdpubsub.ChatStreamNotifyMessage{
-		Status:   string(database.ChatStatusRunning),
-		WorkerID: newWorkerID.String(),
-	}
-	runningPayload, err := json.Marshal(runningNotify)
-	require.NoError(t, err)
-	err = ps.Publish(coderdpubsub.ChatStreamNotifyChannel(chat.ID), runningPayload)
-	require.NoError(t, err)
-
-	// Now release the first dial (if it wasn't already cancelled).
-	close(releaseFirstDial)
-
-	// The subscriber should receive parts from the NEW worker, not the stale one.
-	require.Eventually(t, func() bool {
-		select {
-		case event := <-events:
-			if event.Type == codersdk.ChatStreamEventTypeMessagePart &&
-				event.MessagePart != nil &&
-				event.MessagePart.Part.Text == "new-worker-part" {
-				return true
-			}
-			// If we get the stale part, the bug is present.
-			if event.Type == codersdk.ChatStreamEventTypeMessagePart &&
-				event.MessagePart != nil &&
-				event.MessagePart.Part.Text == "stale-part" {
-				t.Fatal("received stale part from old worker — relay did not cancel in-flight dial")
-			}
-			return false
-		default:
-			return false
-		}
 	}, testutil.WaitMedium, testutil.IntervalFast)
 }
