+++ enterprise/coderd/coderd.go
@@ -191,8 +191,9 @@ func New(ctx context.Context, options *Options) (_ *API, err error) {
 	// This must happen before coderd initialization!
 	options.PostAuthAdditionalHeadersFunc = api.writeEntitlementWarningsHeader
 
-	// Wire up enterprise chat relay for cross-replica message_part streaming.
-	// Must be set before coderd.New so the chat processor gets it.
+	// Wire up enterprise chat subscription with cross-replica relay
+	// and pubsub coordination. Must be set before coderd.New so the
+	// chat processor receives it.
 	replicaHTTPClient := replicaRelayHTTPClient(options.HTTPClient, meshTLSConfig)
 	if replicaHTTPClient == nil {
 		replicaHTTPClient = options.Options.HTTPClient
@@ -200,33 +201,33 @@ func New(ctx context.Context, options *Options) (_ *API, err error) {
 	if replicaHTTPClient == nil {
 		replicaHTTPClient = http.DefaultClient
 	}
-	// Use a closure that captures api by reference so it can access api.AGPL.ID
-	// after coderd.New is called. The provider is only invoked when Subscribe
-	// is called, which happens after initialization, so api.AGPL will be set.
-	options.Options.ChatRemotePartsProvider = func(
-		ctx context.Context,
-		chatID uuid.UUID,
-		workerID uuid.UUID,
-		requestHeader http.Header,
-	) (
-		[]codersdk.ChatStreamEvent,
-		<-chan codersdk.ChatStreamEvent,
-		func(),
-		error,
-	) {
-		// Get the replica ID from the API (will be set after coderd.New)
-		replicaID := api.AGPL.ID
-		if replicaID == uuid.Nil {
-			// Fallback if somehow called before initialization
-			replicaID = uuid.New()
-		}
-		provider := newRemotePartsProvider(
-			resolveReplicaAddress,
-			replicaHTTPClient,
-			replicaID,
-		)
-		return provider(ctx, chatID, workerID, requestHeader)
-	}
+	// Use a closure that captures api by reference so it can access
+	// api.AGPL.ID after coderd.New is called. The SubscribeFn is
+	// only invoked from Subscribe, which happens after init.
+	options.Options.ChatSubscribeFn = newMultiReplicaSubscribeFn(
+		func(
+			ctx context.Context,
+			chatID uuid.UUID,
+			workerID uuid.UUID,
+			requestHeader http.Header,
+		) (
+			[]codersdk.ChatStreamEvent,
+			<-chan codersdk.ChatStreamEvent,
+			func(),
+			error,
+		) {
+			replicaID := api.AGPL.ID
+			if replicaID == uuid.Nil {
+				replicaID = uuid.New()
+			}
+			provider := newRemotePartsProvider(
+				resolveReplicaAddress,
+				replicaHTTPClient,
+				replicaID,
+			)
+			return provider(ctx, chatID, workerID, requestHeader)
+		},
+	)
 
 	api.AGPL = coderd.New(options.Options)
 	defer func() {
